requirements.txt
beautifulsoup4
pandas
loguru
python-dateutil
tqdm
python-dotenv
python-telegram-bot
telegram
requests
lxml
jinja2
sqlitedict
streamlit
ruff